move utilities to src
