+++ src/archive/deploy-registrar.ts
@@ -6,7 +6,7 @@ import {
   DeploymentOutput,
   deploymentsFolder,
   getLogger,
-} from "../../utilities";
+} from "../utilities";
 
 import {
   hashBytecodeWithoutMetadata,
