+++ src/upgradeRegistrarToHaveFilter.ts
@@ -3,7 +3,7 @@ import * as hre from "hardhat";
 import {
   Registrar__factory,
 } from "../typechain";
-import { getLogger } from "../utilities";
+import { getLogger } from "./utilities";
 import { zer0ProtocolAddresses } from "@zero-tech/zero-contracts";
 
 const logger = getLogger("src::test-filter");
