+++ tsconfig.json
@@ -11,6 +11,7 @@
     "resolveJsonModule": true
   },
   "include": [
-    "./src", "./test", "./typechain", "./utilities"],
+    "./src", "./test", "./typechain",
+  ],
   "files": ["./hardhat.config.ts"]
 }
